Test first version with UI for modifier keys
The first version with expandet UI in rebindwidget.ui
for adding modifier keys.

Has same functionality as version befor, but has space for
List of mod keys.

One bug is ecountered in the last branch,
because n-key-rollover with G-keys ist not working correctly.
This is a new bug in the last bugfixing for G-Key detection (macro).